Support shifted add/sub immediates
All immediates are classed as shifted immediates with 0 for unshifted
immediates. This requires a custom operand parser and an alias for
unshifted instructions to add an implicit 0.

Make sure that the call to custom parsers is made from parseOperand()
previously this had been commented out.

Does not yet handle symbolic immediates.